+++ node-proxy/src/router/webui/index.ts
@@ -7,7 +7,7 @@ import { logger } from '@/common/logger'
 import { userInfoMiddleware } from '@/router/webui/middlewares'
 import { encryptFile, searchFile } from '@/utils/convertFile'
 import { encodeFolderName, decodeFolderName } from '@/utils/cryptoUtil'
-import { emptyMiddleware, bodyParserMiddleware } from '@/middleware/common'
+import { emptyMiddleware, bodyParserMiddleware } from '@/utils/middlewares'
 import { RawPasswdInfo, response, splitEncPath } from '@/router/webui/utils'
 import { getUserInfo, cacheUserToken, updateUserInfo } from '@/dao/userDao'
 import { alistServer, webdavServer, port, initAlistConfig, version } from '@/config'
